chore: move filter chip to separate component
