+++ src/dappsExplorer/DAppsExplorerScreenFilter.tsx
@@ -14,7 +14,6 @@ import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context'
 import { useDispatch, useSelector } from 'react-redux'
 import { DappExplorerEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
-import Touchable from 'src/components/Touchable'
 import {
   CategoryWithDapps,
   dappFavoritesEnabledSelector,
@@ -28,6 +27,7 @@ import {
 import { fetchDappsList } from 'src/dapps/slice'
 import { DappFilter, DappSection, DappV1, DappV2 } from 'src/dapps/types'
 import DappCard from 'src/dappsExplorer/DappCard'
+import DappFilterChip from 'src/dappsExplorer/DappFilterChip'
 import FavoriteDappsSection from 'src/dappsExplorer/FavoriteDappsSection'
 import { NoResults } from 'src/dappsExplorer/NoResults'
 import useDappFavoritedToast from 'src/dappsExplorer/useDappFavoritedToast'
@@ -47,13 +47,6 @@ const AnimatedSectionList =
 interface SectionData {
   data: Array<DappV1 | DappV2>
   category: CategoryWithDapps
-}
-
-interface DappFilterChip {
-  chipFilter: DappFilter
-  selectedFilter: DappFilter
-  setFilter: (filter: DappFilter) => void
-  lastChip: boolean
 }
 
 export function DAppsExplorerScreenFilter() {
@@ -270,48 +263,6 @@ function DescriptionView({ message, title }: { message: string; title: string })
   )
 }
 
-function DappFilterChip({ chipFilter, selectedFilter, setFilter, lastChip }: DappFilterChip) {
-  const filterPress = () => {
-    ValoraAnalytics.track(DappExplorerEvents.dapp_filter, { id: chipFilter.id })
-    setFilter(chipFilter)
-  }
-
-  return (
-    <View
-      style={[
-        styles.filterChipContainer,
-        // Filter chips color based on selected filter
-        chipFilter.id === selectedFilter.id
-          ? { backgroundColor: colors.onboardingBlue }
-          : { backgroundColor: colors.onboardingLightBlue },
-        // First Chip has slightly different margins
-        chipFilter.id === 'all'
-          ? { marginLeft: Spacing.Regular16 }
-          : { marginLeft: Spacing.Smallest8 },
-        // Last Chip has slightly different right margin
-        lastChip && { marginRight: Spacing.Regular16 },
-      ]}
-    >
-      <Touchable
-        onPress={filterPress}
-        style={styles.filterChip}
-        testID={`DAppsExplorerScreenFilter/FilterChip/${chipFilter.id}`}
-      >
-        <Text
-          style={[
-            styles.filterChipText,
-            chipFilter.id === selectedFilter.id
-              ? { color: colors.onboardingLightBlue }
-              : { color: colors.onboardingBlue },
-          ]}
-        >
-          {chipFilter.name}
-        </Text>
-      </Touchable>
-    </View>
-  )
-}
-
 const styles = StyleSheet.create({
   safeAreaContainer: {
     flex: 1,
@@ -320,21 +271,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
     flex: 1,
-  },
-  filterChipContainer: {
-    overflow: 'hidden',
-    borderRadius: 94,
-    flex: 1,
-  },
-  filterChip: {
-    minHeight: 32,
-    minWidth: 42,
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: Spacing.Regular16,
-  },
-  filterChipText: {
-    ...fontStyles.small,
   },
   sectionList: {
     flex: 1,
